requirements.txt
python-dotenv
pandas
numpy
scikit-learn
pymongo
certifi
matplotlib
seaborn
pymongo[srv]==3.11

-e .
